pyproject.toml
[tool.poetry]
name = "auto-42share"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
selenium = "3.141.0"
python-dotenv = "^1.0.0"
openpyxl = "^3.1.2"
pyinstaller = "^5.10.1"
PySide6 = "^6.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
